Extract template ans styles into separate files
