+++ ddp-workspace/projects/ddp-sdk/src/lib/components/activityForm/answers/activity-file-answer/activityFileAnswer.component.html
@@ -1,0 +1,27 @@
+<ddp-question-prompt [block]="block"></ddp-question-prompt>
+<div class="file-upload-content">
+    <input type="file" class="file-input"
+           (change)="onFilesSelected($event.target.files)" #fileUpload>
+
+    <div class="file-select">
+        <div class="drop-block" dropFileToUpload (filesDropped)="onFilesSelected($event)">
+            <mat-icon class="upload-icon" color="primary">file_upload</mat-icon>
+            <span class="drop-block-text">Drag files to upload</span>
+        </div>
+        <button mat-raised-button color="primary" class="upload-btn"
+                (click)="fileUpload.click()">Choose file
+            <mat-icon>attach_file</mat-icon>
+        </button>
+
+    </div>
+    <div class="file-info">
+        <span class="file-name">{{file?.name || 'No file uploaded'}}</span>
+    </div>
+</div>
+
+<button mat-raised-button color="primary"
+        class="submit-btn"
+        [disabled]="!isFileSelected"
+        (click)="uploadFile()"
+>Submit
+</button>
