+++ ddp-workspace/projects/ddp-sdk/src/lib/ddp.module.ts
@@ -86,7 +86,7 @@ import { ActivityBooleanAnswer } from './components/activityForm/answers/activit
 import { ActivityAgreementAnswer } from './components/activityForm/answers/activityAgreementAnswer.component';
 import { ActivityTextAnswer } from './components/activityForm/answers/activityTextAnswer.component';
 import { ActivityTextInput } from './components/activityForm/answers/activity-text-input/activityTextInput.component';
-import { ActivityFileAnswer } from './components/activityForm/answers/activityFileAnswer.component';
+import { ActivityFileAnswer } from './components/activityForm/answers/activity-file-answer/activityFileAnswer.component';
 import { ActivityEmailInput } from './components/activityForm/answers/activityEmailInput.component';
 import { ActivityNumericAnswer } from './components/activityForm/answers/activityNumericAnswer.component';
 import { ActivitySectionComponent } from './components/activityForm/activitySection.component';
